backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langgraph
langchain-groq
pinecone
supabase
pydantic
requests
cerebras_cloud_sdk
langchain-cerebras
sentence-transformers
langchain-huggingface
groq
langchain-pinecone
python-multipart
pypdf
edge-tts